refactoring rest layer conans->conan
