+++ conan/internal/rest/file_uploader.py
@@ -3,7 +3,7 @@ import os
 import time
 
 from conan.api.output import ConanOutput, TimedOutput
-from conans.client.rest import response_to_str
+from conan.internal.rest import response_to_str
 from conan.internal.errors import InternalErrorException, RequestErrorException, AuthenticationException, \
     ForbiddenException, NotFoundException
 from conan.errors import ConanException
