+++ conans/server/rest/bottle_routes.py
@@ -1,4 +1,4 @@
-from conans.client.rest.rest_routes import RestRoutes
+from conan.internal.rest.rest_routes import RestRoutes
 
 
 class BottleRoutes(RestRoutes):
