+++ test/integration/configuration/requester_test.py
@@ -5,7 +5,7 @@ import mock
 from mock import Mock, MagicMock
 
 from conan import __version__
-from conans.client.rest.conan_requester import ConanRequester
+from conan.internal.rest.conan_requester import ConanRequester
 from conan.internal.model.conf import ConfDefinition
 from conan.test.utils.tools import temp_folder
 from conans.util.files import save
@@ -21,14 +21,14 @@ class MockRequesterGet(Mock):
 class ConanRequesterCacertPathTests(unittest.TestCase):
     def test_default_no_verify(self):
         mocked_requester = MockRequesterGet()
-        with mock.patch("conans.client.rest.conan_requester.requests", mocked_requester):
+        with mock.patch("conan.internal.rest.conan_requester.requests", mocked_requester):
             requester = ConanRequester(ConfDefinition())
             requester.get(url="aaa", verify=False)
             self.assertEqual(requester._http_requester.verify, False)
 
     def test_default_verify(self):
         mocked_requester = MockRequesterGet()
-        with mock.patch("conans.client.rest.conan_requester.requests", mocked_requester):
+        with mock.patch("conan.internal.rest.conan_requester.requests", mocked_requester):
             requester = ConanRequester(ConfDefinition())
             requester.get(url="aaa", verify=True)
             self.assertEqual(requester._http_requester.verify, True)
@@ -39,7 +39,7 @@ class ConanRequesterCacertPathTests(unittest.TestCase):
         config = ConfDefinition()
         config.update("core.net.http:cacert_path", file_path)
         mocked_requester = MockRequesterGet()
-        with mock.patch("conans.client.rest.conan_requester.requests", mocked_requester):
+        with mock.patch("conan.internal.rest.conan_requester.requests", mocked_requester):
             requester = ConanRequester(config)
             requester.get(url="bbbb", verify=True)
         self.assertEqual(requester._http_requester.verify, file_path)
@@ -48,7 +48,7 @@ class ConanRequesterCacertPathTests(unittest.TestCase):
 class ConanRequesterHeadersTests(unittest.TestCase):
     def test_user_agent(self):
         mock_http_requester = MagicMock()
-        with mock.patch("conans.client.rest.conan_requester.requests", mock_http_requester):
+        with mock.patch("conan.internal.rest.conan_requester.requests", mock_http_requester):
             requester = ConanRequester(ConfDefinition())
             requester.get(url="aaa")
             headers = requester._http_requester.get.call_args[1]["headers"]
