+++ test/integration/graph/version_ranges/version_ranges_cached_test.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 import pytest
 from mock import patch
 
-from conans.client.remote_manager import RemoteManager
+from conan.internal.rest.remote_manager import RemoteManager
 from conan.api.model import RecipeReference
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient, TestServer
