+++ test/integration/remote/retry_test.py
@@ -4,7 +4,7 @@ from collections import namedtuple, Counter
 
 from requests.exceptions import HTTPError
 
-from conans.client.rest.file_uploader import FileUploader
+from conan.internal.rest.file_uploader import FileUploader
 from conan.internal.errors import AuthenticationException, ForbiddenException
 from conan.test.utils.mocks import RedirectedTestOutput
 from conan.test.utils.test_files import temp_folder
