+++ test/integration/tgz_macos_dot_files_test.py
@@ -8,7 +8,7 @@ import unittest
 
 import pytest
 
-from conans.client.remote_manager import uncompress_file
+from conan.internal.rest.remote_manager import uncompress_file
 from conan.api.model import RecipeReference
 from conan.internal.paths import EXPORT_SOURCES_TGZ_NAME
 from conan.test.utils.tools import TestClient, NO_SETTINGS_PACKAGE_ID
